# restful/docs/swagger.yaml
definitions:
...
  models.StoreCart:
    properties:
      addTime:
        description: 添加时间
        type: integer
      bargainId:
        description: 砍价id
        type: integer
      cartNum:
        description: 商品数量
        type: integer
      combinationId:
        description: 拼团id
        type: integer
      id:
        description: 购物车表ID
        type: integer
      isDel:
        description: 是否删除
        type: integer
      isNew:
        description: 是否为立即购买
        type: integer
      isPay:
        description: 0 = 未购买 1 = 已购买
        type: integer
      productAttrUnique:
        description: 商品属性
        type: string
      productId:
        description: 商品ID
        type: integer
      seckillId:
        description: 秒杀产品ID
        type: integer
      type:
        description: 类型
        type: string
      uid:
        description: 用户ID
        type: integer
    type: object
...
  models.StoreCombination:
    properties:
      addTime:
        description: 添加时间
        type: string
      attr:
        description: 活动属性
        type: string
      browse:
        description: 浏览量
        type: integer
      combination:
        type: integer
      cost:
        description: 拼图产品成本
        type: integer
      description:
        description: 拼团内容
        type: string
      id:
        type: integer
      image:
        description: 推荐图
        type: string
      images:
        description: 轮播图
        type: string
      info:
        description: 简介
        type: string
      isDel:
        type: integer
      isHost:
        description: 推荐
        type: integer
      isPostage:
        description: 是否包邮1是0否
        type: integer
      isShow:
        description: 产品状态
        type: integer
      merId:
        description: 商户id
        type: integer
      merUse:
        description: 商户是否可用1可用0不可用
        type: integer
      people:
        description: 参团人数
        type: integer
      postage:
        description: 邮费
        type: number
      price:
        description: 价格
        type: number
      productId:
        description: 商品id
        type: integer
      sales:
        description: 销量
        type: integer
      sort:
        description: 排序
        type: integer
      startTime:
        description: 拼团开始时间
        type: integer
      stock:
        description: 库存
        type: integer
      stopTime:
        description: 拼团结束时间
        type: integer
      title:
        description: 活动标题
        type: string
      unitName:
        description: 单位名
        type: string
    type: object
...
  models.StoreOrder:
    properties:
      addTime:
        description: 创建时间
        type: integer
      backIntegral:
        description: 给用户退了多少积分
        type: number
      bargainId:
        description: 砍价id
        type: integer
      cartId:
        description: 购物车id
        type: string
      combinationId:
        description: 拼团产品id0一般产品
        type: integer
      cost:
        description: 成本价
        type: number
      couponId:
        description: 优惠券id
        type: integer
      couponPrice:
        description: 优惠券金额
        type: number
      deductionPrice:
        description: 抵扣金额
        type: number
      deliveryId:
        description: 快递单号/手机号
        type: string
      deliveryName:
        description: 快递名称/送货人姓名
        type: string
      deliveryType:
        description: 发货类型
        type: string
      gainIntegral:
        description: 消费赚取积分
        type: number
      id:
        description: 订单ID
        type: integer
      isChannel:
        description: 支付渠道(0微信公众号1微信小程序)
        type: integer
      isDel:
        description: 是否删除
        type: integer
      isMerCheck:
        type: integer
      isSystemDel:
        description: 后台管理员删除
        type: integer
      mark:
        description: 备注
        type: string
      merId:
        description: 商户ID
        type: integer
      orderId:
        description: 订单号
        type: string
      paid:
        description: 支付状态
        type: integer
      payPostage:
        description: 支付邮费
        type: number
      payPrice:
        description: 实际支付金额
        type: number
      payTime:
        description: 支付时间
        type: integer
      payType:
        description: 支付方式
        type: string
      pinkId:
        description: 拼团id 0没有拼团
        type: integer
      realName:
        description: 用户姓名
        type: string
      refundPrice:
        description: 退款金额
        type: number
      refundReason:
        description: 不退款的理由
        type: string
      refundReasonTime:
        description: 退款时间
        type: integer
      refundReasonWap:
        description: 前台退款原因
        type: string
      refundReasonWapExplain:
        description: 退款用户说明
        type: string
      refundReasonWapImg:
        description: 退款图片
        type: string
      refundStatus:
        description: 0 未退款 1 申请中 2 已退款
        type: integer
      remark:
        description: 管理员备注
        type: string
      seckillId:
        description: 秒杀产品ID
        type: integer
      status:
        description: '订单状态（-1 : 申请退款 -2 : 退货成功 0：待发货；1：待收货；2：已收货；3：待评价；-1：已退款）'
        type: integer
      totalNum:
        description: 订单商品总数
        type: integer
      totalPostage:
        description: 邮费
        type: number
      totalPrice:
        description: 订单总价
        type: number
      uid:
        description: 用户id
        type: integer
      unique:
        description: 唯一id(md5加密)类似id
        type: string
      useIntegral:
        description: 使用积分
        type: number
      userAddress:
        description: 详细地址
        type: string
      userPhone:
        description: 用户电话
        type: string
    type: object
...
info:
  contact: {}
  license: {}
paths:
...
  /storeCarts:
    get:
      parameters:
      - description: query, ?query=age:>:21,name:like:%jason%
        in: query
        name: query
        type: string
      - description: order, ?order=age:desc,created_at:asc
        in: query
        name: order
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: storeCart array
          schema:
            items:
              $ref: '#/definitions/models.StoreCart'
            type: array
      summary: List
      tags:
      - StoreCart
    post:
      parameters:
      - description: StoreCart
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.StoreCart'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create
      tags:
      - StoreCart
  /storeCarts/{storeCartId}:
    delete:
      parameters:
      - description: storeCartId
        in: path
        name: storeCartId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete
      tags:
      - StoreCart
    get:
      parameters:
      - description: storeCartId
        in: path
        name: storeCartId
        required: true
        type: string
      responses:
        "200":
          description: storeCart object
          schema:
            $ref: '#/definitions/models.StoreCart'
      summary: Get
      tags:
      - StoreCart
    patch:
      parameters:
      - description: storeCart
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.StoreCart'
      - description: storeCartId
        in: path
        name: storeCartId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Patch
      tags:
      - StoreCart
    put:
      parameters:
      - description: storeCart
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.StoreCart'
      - description: storeCartId
        in: path
        name: storeCartId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Put
      tags:
      - StoreCart
...
  /storeCombinations:
    get:
      parameters:
      - description: query, ?query=age:>:21,name:like:%jason%
        in: query
        name: query
        type: string
      - description: order, ?order=age:desc,created_at:asc
        in: query
        name: order
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: storeCombination array
          schema:
            items:
              $ref: '#/definitions/models.StoreCombination'
            type: array
      summary: List
      tags:
      - StoreCombination
    post:
      parameters:
      - description: StoreCombination
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.StoreCombination'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create
      tags:
      - StoreCombination
  /storeCombinations/{storeCombinationId}:
    delete:
      parameters:
      - description: storeCombinationId
        in: path
        name: storeCombinationId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete
      tags:
      - StoreCombination
    get:
      parameters:
      - description: storeCombinationId
        in: path
        name: storeCombinationId
        required: true
        type: string
      responses:
        "200":
          description: storeCombination object
          schema:
            $ref: '#/definitions/models.StoreCombination'
      summary: Get
      tags:
      - StoreCombination
    patch:
      parameters:
      - description: storeCombination
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.StoreCombination'
      - description: storeCombinationId
        in: path
        name: storeCombinationId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Patch
      tags:
      - StoreCombination
    put:
      parameters:
      - description: storeCombination
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.StoreCombination'
      - description: storeCombinationId
        in: path
        name: storeCombinationId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Put
      tags:
      - StoreCombination
...
  /storeOrders:
    get:
      parameters:
      - description: query, ?query=age:>:21,name:like:%jason%
        in: query
        name: query
        type: string
      - description: order, ?order=age:desc,created_at:asc
        in: query
        name: order
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: storeOrder array
          schema:
            items:
              $ref: '#/definitions/models.StoreOrder'
            type: array
      summary: List
      tags:
      - StoreOrder
    post:
      parameters:
      - description: StoreOrder
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.StoreOrder'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create
      tags:
      - StoreOrder
  /storeOrders/{storeOrderId}:
    delete:
      parameters:
      - description: storeOrderId
        in: path
        name: storeOrderId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete
      tags:
      - StoreOrder
    get:
      parameters:
      - description: storeOrderId
        in: path
        name: storeOrderId
        required: true
        type: string
      responses:
        "200":
          description: storeOrder object
          schema:
            $ref: '#/definitions/models.StoreOrder'
      summary: Get
      tags:
      - StoreOrder
    patch:
      parameters:
      - description: storeOrder
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.StoreOrder'
      - description: storeOrderId
        in: path
        name: storeOrderId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Patch
      tags:
      - StoreOrder
    put:
      parameters:
      - description: storeOrder
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.StoreOrder'
      - description: storeOrderId
        in: path
        name: storeOrderId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Put
      tags:
      - StoreOrder
...
swagger: "2.0"